api-importer/requirements.txt
cheroot==10.0.1
Flask==3.1.2
flask_restx==1.3.0
PyYAML==6.0.2
Requests==2.32.5
